Bugfix: add missing column 'can_manage_roles' in public.v_users_self
I originally intended to only provide the necessary columns.
However, in an attempt to simplify the PHP end of this, I decided to add NULL columns for such cases.
When I did this, I overlooked the 'can_manage_roles' column.

Looking at this, I might also be better of setting the values to FALSE instead of null::bool where is can be easily determined to be FALSE.