Cleanup: restructure user dashboard, login, and logout links to be more consistent with the 'user_*' naming
Having all the dashboards called just 'dashboard' is convenient in the general sense, but is more inconsistent when I have a notable amount of 'user_*' names.
Be consistent with the 'user_*' names and have the login, logout, and (user) dashboard files renamed to user_login, user_logout, and user_dashboard.

The other users of 'dashboard' will be updated as needed and as I get to them.