Progress: User Profile / User View Page
The profile images should be file ids and not byteas.
- These columns should also be foreign keys to the file table.

Enable use of profile image in the user object.

Update/Fix the Japanese related translations.

The User Profile / User View page is intended to be a simpler version of the User Settings page with a focus on user information and less focus on user settings values.