Bugfix: set_error() must accept an array of errors
This is considered a bug because the current design of get_error() returns an array if errors.
Prior to this, calling set_error($some_class->get_error()) would result in no error being reported because the array value gets ignored.

With this change, the entire array contents are appended, only if the given array value is a valid error class instance.