Progress: begin adding query helper classes and traits
These are some traits and classes used to help facilitate the generation of SQL queries according to the postgresql standard.
Because of the volume of work being done, I am committing the very much incomplete and incorrect work.
A number of the files are copy and pasted placeholders until I can get to actually coding it and will contain invalid content until such time I can complete this feature set.
There may be futher restructing.