Progress: continue development of database abstraction
I am having a lot of indecision on how exactly to approach key parts of this.
I have added query parameter class, and may be replacing the existing implementation (or not).
I decided to keep the naming system more simple by remove the query_ prefix of methods and variables provided by the database traits.
This may introduce naming conflicts that need to be reviewed and resolved.