Update: rename 'database_option' to 'database_cascade', and cleanup relating code
The use of 'option' is not quite accurate, so I renamed it to the more aptly named 'cascade'.
The constraint code that conditionally utilized cascade has been rewritten to be stored in a single variable (as an array).
There is no cascade trait like there was an option trait because its use is different from case to case.