Feature: t_base_return class objects may now be used directly as strings
By utilizing __toString(), all classes using t_base_return can now be directly used as a string.
This saves the effort of writing $my_class->get_value() or $my_class->get_value_exact().

This directly returns the private value as a string, which is probably faster than returning an object and then calling a function to convert.

Complex types, such as array, object, or resource, or converted to an empty string and must still be handled via function calls.